renaming to entity package
