+++ src/main/java/com/hftamayo/java/todo/services/impl/RolesServiceImpl.java
@@ -3,8 +3,8 @@ package com.hftamayo.java.todo.services.impl;
 import com.hftamayo.java.todo.dao.RolesDao;
 import com.hftamayo.java.todo.dto.roles.RolesResponseDto;
 import com.hftamayo.java.todo.exceptions.EntityAlreadyExistsException;
-import com.hftamayo.java.todo.model.ERole;
-import com.hftamayo.java.todo.model.Roles;
+import com.hftamayo.java.todo.entity.ERole;
+import com.hftamayo.java.todo.entity.Roles;
 import com.hftamayo.java.todo.services.RolesService;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
