+++ src/main/java/com/hftamayo/java/todo/services/impl/UserServiceImpl.java
@@ -3,8 +3,8 @@ package com.hftamayo.java.todo.services.impl;
 import com.hftamayo.java.todo.dao.UserDao;
 import com.hftamayo.java.todo.dao.RolesDao;
 import com.hftamayo.java.todo.dto.user.UserResponseDto;
-import com.hftamayo.java.todo.model.Roles;
-import com.hftamayo.java.todo.model.User;
+import com.hftamayo.java.todo.entity.Roles;
+import com.hftamayo.java.todo.entity.User;
 import com.hftamayo.java.todo.services.UserService;
 import com.hftamayo.java.todo.exceptions.EntityAlreadyExistsException;
 import lombok.RequiredArgsConstructor;
